+++ R/lasso-deployment.R
@@ -5,6 +5,7 @@
 #' \item Load a saved model from \code{\link{LassoDevelopment}}
 #' \item Run the model against test data to generate predictions
 #' \item Push these predictions to SQL Server
+#' \item Identify factors that could benefit outcomes (see final examples)
 #' }
 #' @docType class
 #' @usage LassoDeployment(type, df, grainCol, predictedCol, impute, debug, cores, modelName)
@@ -17,9 +18,9 @@
 #' much as possible. Number of columns, names, types, grain, and predicted must be the same.
 #' @param grainCol The dataframe's column that has IDs pertaining to the grain
 #' @param predictedCol Column that you want to predict.
-#' @param impute For training df, set all-column imputation to T or F.
-#' If T, this uses values calculated in development.
-#' F leads to removal of rows containing NULLs and is not recommended.
+#' @param impute For training df, set all-column imputation to TRUE or FALSE.
+#' If TRUE, this uses values calculated in development.
+#' FALSE leads to removal of rows containing NULLs and is not recommended.
 #' @param debug Provides the user extended output to the console, in order
 #' to monitor the calculations throughout. Use T or F.
 #' @param cores Number of cores you'd like to use. Defaults to 2.
@@ -39,16 +40,43 @@
 #' Return the grain, all top factors, and their weights. \cr
 #' \emph{Usage:} \code{$getTopFactors(numberOfFactors = NA, includeWeights = FALSE)} \cr
 #' Params: \cr
-#'   - \code{numberOfFactors:} returns the top \code{n} factors. Defaults to all factors. \cr
-#'   - \code{includeWeights:} If \code{TRUE}, returns weights associated with each factor.
+#'   - \code{numberOfFactors} returns the top \code{n} factors. Defaults to all factors. \cr
+#'   - \code{includeWeights} If \code{TRUE}, returns weights associated with each factor.
 #' @section \code{$getOutDf()}:
 #' Returns the output dataframe. \cr
 #' \emph{Usage:} \code{$getOutDf()} 
+#' @section \code{$performNewPredictions()}:
+#' Performs predictions on new data which has gone through the necessary
+#' preprocessing. \cr
+#' \emph{Usage:} \code{$performNewPredictions(newData)} \cr
+#' Params: \cr
+#'   - \code{newData} A dataframe on which the model can make predictions.
+#' @section \code{$getProcessVariablesDf()}:
+#' Get a data frame with expected changes in outcomes conditional on process
+#' variable changes. \cr
+#' \emph{Usage:} \code{$getProcessVariablesDf(modifiableVariables, 
+#' variableLevels = NULL, grainColumnValues = NULL, smallerBetter = TRUE, 
+#' repeatedFactors = FALSE, numTopFactors = 3)} \cr
+#' Params: \cr
+#'   - \code{modifiableVariables} A vector of names of modifiable variables.\cr
+#'   - \code{variableLevels} A list of variable values indexed by 
+#'   modifiable variable names. This allows one to use numeric variables by 
+#'   specifying baselines and to restrict categorical variables by limiting 
+#'   which factors can be surfaced.\cr
+#'   - \code{grainColumnIDs} A vector of grain column IDs. If \code{NULL}, the 
+#'   whole deployment dataframe will be used.\cr
+#'   - \code{smallerBetter} A boolean determining whether or not lower 
+#'   predictions/probabilities are more desirable. \cr
+#'   - \code{repeatedFactors} A boolean determining whether or not a single 
+#'   modifiable factor can be listed several times. \cr
+#'   - \code{numTopFactors} The number of modifiable process variables to 
+#'   include in each row.
 #' @export
 #' @seealso \code{\link{healthcareai}}
 #' @seealso \code{\link{writeData}}
 #' @seealso \code{\link{selectData}}
 #' @examples
+#' 
 #' 
 #' #### Classification Example using csv data ####
 #' ## 1. Loading data and packages.
@@ -403,6 +431,103 @@
 #'           tableName = 'HCRDeployRegressionBASE')
 #' 
 #' print(proc.time() - ptm)
+#' 
+#' #### Identify factors that could benefit outcomes: getProcessVariablesDf ####
+#' #############################################################################
+#' 
+#' # getProcessVariableDf() identifies opportunities for improved outcomes at
+#' # the grain level. It is important that the variables ("modifiableVariables")
+#' # and values ("variableLevels") used in this function are under the control 
+#' # of the care management process. The best use case for this function is a
+#' # "natural experiment" where otherwise similar groups had different 
+#' # treatments applied to them, and that treatment is the modifiable variable
+#' # of interest.
+#'
+#' # This example shows how to use the getProcessVariableDf() function, using 
+#' # another readmission-prediction model. In this example systolic blood pressure 
+#' # is converted into a categorical variable to demonstrate functionality.
+#' # Because of the lasso's automatic feature selection, this example is fairly
+#' # limited. For a wider variety of examples, see ?RandomForestDeployment
+#' 
+#' csvfile <- system.file("extdata", 
+#'                        "HCRDiabetesClinical.csv", 
+#'                        package = "healthcareai")
+#' 
+#' # Replace csvfile with 'path/file'
+#' df <- read.csv(file = csvfile, 
+#'                header = TRUE, 
+#'                na.strings = c("NULL", "NA", ""))
+#' 
+#' df$PatientID <- NULL # Remove extra ID
+#' 
+#' # Save a dataframe for validation later on
+#' dfDeploy <- df[951:1000,]
+#' 
+#' ## Develop and Deploy the model
+#' set.seed(42)
+#' p <- SupervisedModelDevelopmentParams$new()
+#' p$df <- df
+#' p$type <- "classification"
+#' p$impute <- TRUE
+#' p$grainCol <- "PatientEncounterID"
+#' p$predictedCol <- "ThirtyDayReadmitFLG"
+#' p$debug <- FALSE
+#' p$cores <- 1
+#' 
+#' Lasso <- LassoDevelopment$new(p)
+#' Lasso$run()
+#' 
+#' p2 <- SupervisedModelDeploymentParams$new()
+#' p2$type <- "classification"
+#' p2$df <- dfDeploy
+#' p2$grainCol <- "PatientEncounterID"
+#' p2$predictedCol <- "ThirtyDayReadmitFLG"
+#' p2$impute <- TRUE
+#' p2$debug <- FALSE
+#' p2$cores <- 1
+#' 
+#' dL <- LassoDeployment$new(p2)
+#' dL$deploy()
+#' 
+#' ## Get predicted outcome changes using getProcessVariablesDf
+#' 
+#' # getProcessVariablesDf only uses variables with non-zero coefficients, 
+#' # automatically discarding the rest. In this example, only A1CNBR had a
+#' # nonzerocoefficient. Even though this variable is numeric, we can still use
+#' # getProcessVariablesDf using the variableLevels parameter. For examples with
+#' # categorical variables and additional details about getProcessVariablesDf,
+#' # see the examples in ?RandomForestDeployment
+#' 
+#' dL$getProcessVariablesDf(modifiableVariables = c("A1CNBR"),
+#'                          variableLevels = list(A1CNBR = c(5.6, 6.0, 6.5)))
+#' 
+#' # By default, the function returns predictions for all rows, but we can 
+#' # restrict to specific rows using the grainColumnIDs parameter
+#' dL$getProcessVariablesDf(modifiableVariables = c("A1CNBR"),
+#'                          variableLevels = list(A1CNBR = c(5.6, 6.0, 6.5)), 
+#'                          grainColumnIDs = c(951, 975))
+#' 
+#' # The repeatedFactors parameter allows one to get multiple predictions  
+#' # forthe same variable. For example, reducing A1C to 5.6 might most improve a 
+#' # patient's risk, but reducing A1C to 5.9 is likely to also reduce the risk.
+#' dL$getProcessVariablesDf(modifiableVariables = c("A1CNBR"),
+#'                          variableLevels = list(A1CNBR = c(5.6, 5.9, 6.2, 6.5)),
+#'                          repeatedFactors = TRUE)
+#' 
+#' # The numTopFactors parameter allows one to set the maximum number of 
+#' # predictions to display (with the default being 3)
+#' dL$getProcessVariablesDf(modifiableVariables = c("A1CNBR"),
+#'                          variableLevels = list(A1CNBR = c(5.6, 5.9, 6.2, 6.5)),
+#'                          repeatedFactors = TRUE, 
+#'                          numTopFactors = 2)
+#' 
+#' # If we want to make predictions for increasing the probability (not
+#' # likely in the case of readmissions), we can do so using the smallerBetter
+#' # parameter. (Here, all the deltas will be non-negative, corresponding to an 
+#' # increased risk)
+#' dL$getProcessVariablesDf(modifiableVariables = c("A1CNBR"),
+#'                          variableLevels = list(A1CNBR = c(5.6, 6.0, 6.5)),
+#'                          smallerBetter = FALSE)
 
 LassoDeployment <- R6Class(
   "LassoDeployment",
@@ -444,10 +569,7 @@ LassoDeployment <- R6Class(
       private$lambda1se <- private$fitGrLasso$lambda[private$indLambda1se]
       
       # Predictions (in terms of probability)
-      private$predictions <- stats::predict(object = private$fitGrLasso,
-                                     X = model.matrix(private$modFmla, data = self$params$df)[,-1],
-                                     lambda = private$lambda1se,
-                                     type = "response")
+      private$predictions <- self$performNewPredictions(self$params$df)
       
       if (isTRUE(self$params$debug)) {
         cat("Rows in prob prediction: ", nrow(private$predictedVals), '\n')
@@ -519,6 +641,9 @@ LassoDeployment <- R6Class(
 
     #Override: deploy the model
     deploy = function() {
+      
+      # Start sink to capture console ouptut
+      sink("tmp_prediction_console_output.txt", append = FALSE, split = TRUE)
 
       # Try to load the model
       private$fitGrLasso <- private$fitObj
@@ -548,11 +673,24 @@ LassoDeployment <- R6Class(
 
       # create dataframe for output
       super$createDf()
+
+      sink()  # Close connection
+      # Get metadata, attach to output DF and write to text file
+      super$getMetadata()
+
     },
     
     # Surface outDf as attribute for export to Oracle, MySQL, etc
     getOutDf = function() {
       return(private$outDf)
+    }, 
+    
+    # Perform predictions on new data
+    performNewPredictions = function(newData) {
+      stats::predict(object = private$fitGrLasso,
+                     X = model.matrix(private$modFmla, data = newData)[,-1],
+                     lambda = private$lambda1se,
+                     type = "response")
     }
   )
 )